# Add optimised HasFlagFast<T> extension to Enum