--- a/osu.Framework/Extensions/EnumExtensions/EnumExtensions.cs
+++ b/osu.Framework/Extensions/EnumExtensions/EnumExtensions.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using osu.Framework.Utils;
 
 namespace osu.Framework.Extensions.EnumExtensions
@@ -50,6 +51,47 @@ namespace osu.Framework.Extensions.EnumExtensions
 
                 throw new ArgumentException($"Not all values of {typeof(T)} have {nameof(OrderAttribute)} specified.");
             });
+        }
+
+        /// <summary>
+        /// A fast alternative functionally equivalent to <see cref="Enum.HasFlag"/>, eliminating boxing in all scenarios.
+        /// </summary>
+        /// <param name="enumValue">The enum to check.</param>
+        /// <param name="flag">The flag to check for.</param>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static unsafe bool HasFlagFast<T>(this T enumValue, T flag) where T : unmanaged, Enum
+        {
+            // Note: Using a switch statement would eliminate inlining.
+
+            if (sizeof(T) == 1)
+            {
+                var value1 = Unsafe.As<T, byte>(ref enumValue);
+                var value2 = Unsafe.As<T, byte>(ref flag);
+                return (value1 & value2) > 0;
+            }
+
+            if (sizeof(T) == 2)
+            {
+                var value1 = Unsafe.As<T, short>(ref enumValue);
+                var value2 = Unsafe.As<T, short>(ref flag);
+                return (value1 & value2) > 0;
+            }
+
+            if (sizeof(T) == 4)
+            {
+                var value1 = Unsafe.As<T, int>(ref enumValue);
+                var value2 = Unsafe.As<T, int>(ref flag);
+                return (value1 & value2) > 0;
+            }
+
+            if (sizeof(T) == 8)
+            {
+                var value1 = Unsafe.As<T, long>(ref enumValue);
+                var value2 = Unsafe.As<T, long>(ref flag);
+                return (value1 & value2) > 0;
+            }
+
+            throw new ArgumentException($"Invalid enum type provided: {typeof(T)}.");
         }
     }
 }