--- a/osu.Framework/Graphics/Containers/ContainerExtensions.cs
+++ b/osu.Framework/Graphics/Containers/ContainerExtensions.cs
@@ -4,6 +4,7 @@
 using osuTK;
 using System;
 using System.Collections.Generic;
+using osu.Framework.Extensions.EnumExtensions;
 
 namespace osu.Framework.Graphics.Containers
 {
@@ -43,13 +44,13 @@ namespace osu.Framework.Graphics.Containers
 
             // For anchor/origin positioning to be preserved correctly,
             // relatively sized axes must be lifted to the wrapping container.
-            if (container.RelativeSizeAxes.HasFlag(Axes.X))
+            if (container.RelativeSizeAxes.HasFlagFast(Axes.X))
             {
                 container.Width = drawable.Width;
                 drawable.Width = 1;
             }
 
-            if (container.RelativeSizeAxes.HasFlag(Axes.Y))
+            if (container.RelativeSizeAxes.HasFlagFast(Axes.Y))
             {
                 container.Height = drawable.Height;
                 drawable.Height = 1;