--- a/osu.Framework/Graphics/Containers/FillFlowContainer.cs
+++ b/osu.Framework/Graphics/Containers/FillFlowContainer.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using osuTK;
 using System.Linq;
+using osu.Framework.Extensions.EnumExtensions;
 using osu.Framework.Utils;
 
 namespace osu.Framework.Graphics.Containers
@@ -87,9 +88,9 @@ namespace osu.Framework.Graphics.Containers
         private Vector2 spacingFactor(Drawable c)
         {
             Vector2 result = c.RelativeOriginPosition;
-            if (c.Anchor.HasFlag(Anchor.x2))
+            if (c.Anchor.HasFlagFast(Anchor.x2))
                 result.X = 1 - result.X;
-            if (c.Anchor.HasFlag(Anchor.y2))
+            if (c.Anchor.HasFlagFast(Anchor.y2))
                 result.Y = 1 - result.Y;
             return result;
         }
@@ -104,8 +105,8 @@ namespace osu.Framework.Graphics.Containers
 
                 // If we are autosize and haven't specified a maximum size, we should allow infinite expansion.
                 // If we are inheriting then we need to use the parent size (our ActualSize).
-                max.X = AutoSizeAxes.HasFlag(Axes.X) ? float.MaxValue : s.X;
-                max.Y = AutoSizeAxes.HasFlag(Axes.Y) ? float.MaxValue : s.Y;
+                max.X = AutoSizeAxes.HasFlagFast(Axes.X) ? float.MaxValue : s.X;
+                max.Y = AutoSizeAxes.HasFlagFast(Axes.Y) ? float.MaxValue : s.Y;
             }
 
             var children = FlowingChildren.ToArray();
@@ -262,17 +263,17 @@ namespace osu.Framework.Graphics.Containers
                         break;
                 }
 
-                if (c.Anchor.HasFlag(Anchor.x1))
+                if (c.Anchor.HasFlagFast(Anchor.x1))
                     // Begin flow at centre of row
                     result[i].X += rowOffsetsToMiddle[rowIndices[i]];
-                else if (c.Anchor.HasFlag(Anchor.x2))
+                else if (c.Anchor.HasFlagFast(Anchor.x2))
                     // Flow right-to-left
                     result[i].X = -result[i].X;
 
-                if (c.Anchor.HasFlag(Anchor.y1))
+                if (c.Anchor.HasFlagFast(Anchor.y1))
                     // Begin flow at centre of total height
                     result[i].Y -= height / 2;
-                else if (c.Anchor.HasFlag(Anchor.y2))
+                else if (c.Anchor.HasFlagFast(Anchor.y2))
                     // Flow bottom-to-top
                     result[i].Y = -result[i].Y;
             }