--- a/osu.Framework/Graphics/Containers/TextFlowContainer.cs
+++ b/osu.Framework/Graphics/Containers/TextFlowContainer.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using osu.Framework.Extensions.EnumExtensions;
 
 namespace osu.Framework.Graphics.Containers
 {
@@ -192,7 +193,7 @@ namespace osu.Framework.Graphics.Containers
         {
             // FillFlowContainer will reverse the ordering of right-anchored words such that the (previously) first word would be
             // the right-most word, whereas it should still be flowed left-to-right. This is achieved by reversing the comparator.
-            if (TextAnchor.HasFlag(Anchor.x2))
+            if (TextAnchor.HasFlagFast(Anchor.x2))
                 return base.Compare(y, x);
 
             return base.Compare(x, y);