--- a/osu.Framework/Graphics/Lines/Path.cs
+++ b/osu.Framework/Graphics/Lines/Path.cs
@@ -9,6 +9,7 @@ using osu.Framework.Graphics.Shaders;
 using osu.Framework.Allocation;
 using System.Collections.Generic;
 using osu.Framework.Caching;
+using osu.Framework.Extensions.EnumExtensions;
 using osuTK.Graphics;
 using osuTK.Graphics.ES30;
 
@@ -113,7 +114,7 @@ namespace osu.Framework.Graphics.Lines
         {
             get
             {
-                if (AutoSizeAxes.HasFlag(Axes.X))
+                if (AutoSizeAxes.HasFlagFast(Axes.X))
                     return base.Width = vertexBounds.Width;
 
                 return base.Width;
@@ -131,7 +132,7 @@ namespace osu.Framework.Graphics.Lines
         {
             get
             {
-                if (AutoSizeAxes.HasFlag(Axes.Y))
+                if (AutoSizeAxes.HasFlagFast(Axes.Y))
                     return base.Height = vertexBounds.Height;
 
                 return base.Height;