… CodeAnalysis/BannedSymbols.txt
… osu.Framework/Graphics/UserInterface/Menu.cs
… osu.Framework/Graphics/UserInterface/TabControl.cs
… osu.Framework/Input/Handlers/Mouse/OsuTKRawMouseHandler.cs
… osu.Framework/Input/UserInputManager.cs
… osu.Framework/Platform/MacOS/OsuTKMacOSWindow.cs
… osu.Framework/Platform/OsuTKWindow.cs
… osu.Framework/Platform/SDL2/SDL2Extensions.cs
… osu.Framework/Platform/SDL2DesktopWindow.cs