@peppy 2021-02-25
Is it worth adding the comparison benchmark to the benchmarks project?